separate yml files and clean up service names following namespace convenion
