+++ src/PrestaShopBundle/Resources/config/services/bundle/bridge.yml
@@ -1,103 +1,25 @@
+imports:
+  - { resource: bridge/*.yml }
+
 services:
   _defaults:
     public: true
 
-  # @todo: services naming should follow classnames
-  prestashop.core.bridge.smarty_bridge:
-    class: PrestaShopBundle\Bridge\Smarty\SmartyBridge
-    arguments:
-      - "@prestashop.adapter.legacy.context"
-      - "@prestashop.adapter.legacy.configuration"
-      - !tagged core.bridge.smarty_configurator
-
-  prestashop.core.bridge.controller_configuration_factory:
+  prestashop.bridge.admin_controller.controller_configuration_factory:
     class: PrestaShopBundle\Bridge\AdminController\ControllerConfigurationFactory
     arguments:
       - "@prestashop.user_provider"
 
-  prestashop.core.bridge.admin_controller.legacy_controller_bridge_factory:
+  prestashop.bridge.admin_controller.legacy_controller_bridge_factory:
     class: PrestaShopBundle\Bridge\AdminController\LegacyControllerBridgeFactory
     arguments:
-      - "@prestashop.core.bridge.controller_configuration_factory"
+      - "@prestashop.bridge.admin_controller.controller_configuration_factory"
       - "@prestashop.adapter.multistore_feature"
 
-  prestashop.core.bridge.breadcrumbs_and_title_configurator:
-    class: PrestaShopBundle\Bridge\Smarty\BreadcrumbsAndTitleConfigurator
-    arguments:
-      - "@prestashop.adapter.legacy.context"
-      - "@prestashop.adapter.legacy.configuration"
-    tags: [ 'core.bridge.smarty_configurator' ]
-
-  prestashop.core.bridge.footer_configurator:
-    class: PrestaShopBundle\Bridge\Smarty\FooterConfigurator
-    arguments:
-      - "@prestashop.adapter.legacy.context"
-    tags: [ 'core.bridge.smarty_configurator' ]
-
-  prestashop.core.bridge.header_configurator:
-    class: PrestaShopBundle\Bridge\Smarty\HeaderConfigurator
-    arguments:
-      - "@prestashop.router"
-      - "@translator"
-      - "@prestashop.adapter.legacy.context"
-      - "@prestashop.hook.dispatcher"
-      - "@prestashop.adapter.legacy.configuration"
-    tags: [ 'core.bridge.smarty_configurator' ]
-
-  prestashop.core.bridge.modal_configurator:
-    class: PrestaShopBundle\Bridge\Smarty\ModalConfigurator
-    tags: [ 'core.bridge.smarty_configurator' ]
-
-  prestashop.core.bridge.notification_configurator:
-    class: PrestaShopBundle\Bridge\Smarty\NotificationsConfigurator
-    arguments:
-      - "@prestashop.adapter.legacy.configuration"
-    tags: [ 'core.bridge.smarty_configurator' ]
-
-  prestashop.core.bridge.toolbar_flags_configurator:
-    class: PrestaShopBundle\Bridge\Smarty\ToolbarFlagsConfigurator
-    arguments:
-      - "@prestashop.router"
-      - "@prestashop.adapter.legacy.configuration"
-    tags: [ 'core.bridge.smarty_configurator' ]
-
-  prestashop.core.bridge.helper_list_bridge:
-    class: PrestaShopBundle\Bridge\Helper\HelperListBridge
-    arguments:
-      - "@prestashop.adapter.legacy.context"
-      - "@prestashop.user_provider"
-      - "@prestashop.core.bridge.helper_list_configurator"
-      - "@prestashop.hook.dispatcher"
-      - "@prestashop.adapter.legacy.configuration"
-
-  prestashop.core.bridge.helper_list_configuration_factory:
-    class: PrestaShopBundle\Bridge\Helper\HelperListConfigurationFactory
-
-  prestashop.core.bridge.helper_list_configurator:
-    class: PrestaShopBundle\Bridge\Helper\HelperListConfigurator
-    arguments:
-      - "@prestashop.core.bridge.breadcrumbs_and_title_configurator"
-
-  prestashop.core.bridge.processor.filters_helper:
-    class: PrestaShopBundle\Bridge\Helper\FiltersHelper
-    arguments:
-      - "@prestashop.adapter.legacy.context"
-      - "@prestashop.hook.dispatcher"
-
-  prestashop.core.bridge.processor.reset_filters_helper:
-    class: PrestaShopBundle\Bridge\Helper\ResetFiltersHelper
-    arguments:
-      - "@prestashop.adapter.legacy.context"
-
-  PrestaShopBundle\Bridge\AdminController\Listener\InitFrameworkBridgeControllerListener:
+  prestashop.bridge.admin_controller.listener.init_framework_bridge_controller_listener:
     class: PrestaShopBundle\Bridge\AdminController\Listener\InitFrameworkBridgeControllerListener
     tags:
       - { name: kernel.event_listener, event: kernel.controller }
     arguments:
       - "@prestashop.adapter.legacy.context"
       - "@prestashop.core.localization.locale.repository"
-
-  prestashop.core.bridge.helper_list_feature:
-    class: PrestaShopBundle\Bridge\Helper\HelperListCustomizer\HelperListFeatureBridge
-    parent: prestashop.core.bridge.helper_list_bridge
-    public: true
