+++ src/PrestaShopBundle/Resources/config/services/bundle/bridge/helper/_list_customizer.yml
@@ -1,0 +1,8 @@
+services:
+  _defaults:
+    public: true
+
+  prestashop.bridge.helper.list_customizer.feature_helper_list_bridge:
+    class: PrestaShopBundle\Bridge\Helper\ListCustomizer\FeatureHelperListBridge
+    parent: prestashop.bridge.helper.helper_list_bridge
+    public: true
